requirements.txt
torch 
torchvision
torchaudio
transformers
opencv-python 
Pillow 
tqdm  
nest-asyncio
psutil
pyzmq
python-dateutil